[Patchset 5] Fix chromium build
First failure: https://chromium-review.googlesource.com/c/chromium/src/+/1480662
Looks caused by uint vs. int comparison in ASSERT_EQ.

TBR=primiano@google.com

+++ src/base/circular_queue_unittest.cc
@@ -70,7 +70,7 @@ TEST(CircularQueueTest, Int) {
   {
     std::minstd_rand0 rnd_engine(0);
     it = queue.begin();
-    for (int i = 0; i < kNumInts; ++i, ++it) {
+    for (size_t i = 0; i < kNumInts; ++i, ++it) {
       ASSERT_LT(it, queue.end());
       ASSERT_EQ(*it, static_cast<int>(rnd_engine()));
     }
